pyapi/shared/requirements-test.txt
coverage>=3.7.1
flake8>=2.4.1
mock>=1.0.1
nose>=1.3.7
pbr>=1.2.0
pylint>=1.4.4
pyramid>=1.6a2
python-keystoneclient==1.6.0
python-swiftclient==2.4.0
PyYAML>=3.11
responses>=0.4.0
tissue>=0.9.2
tox>=2.1.1
waitress>=0.8.9
webtest>=2.0.18
